Updated frontend for precision change

+++ GUI/MandelbrotSetVisualizer/src-tauri/OpenCLKernel.cl
@@ -1,6 +1,6 @@
 typedef struct {
-	double real;
-	double imag;
+	float real;
+	float imag;
 } Complex;
 
 __kernel void calculateIters(__global Complex* IN, __global int* OUT, const unsigned int max_iter)
@@ -8,14 +8,14 @@ __kernel void calculateIters(__global Complex* IN, __global int* OUT, const unsi
 	int idx = get_global_id(0);
 	Complex c = IN[idx];
 	
-	double x0 = c.real;
-	double y0 = c.imag;
+	float x0 = c.real;
+	float y0 = c.imag;
 
-	double x2 = 0;
-	double y2 = 0;
+	float x2 = 0;
+	float y2 = 0;
 
-	double x = 0;
-	double y = 0;
+	float x = 0;
+	float y = 0;
 	
 	int result = -1;
 	for (int i = 0; i < max_iter; i++) {
